user-service/requirements.txt
# flask==2.1.3
# gunicorn==20.1.0
# requests==2.31.0

Flask==2.2.5
Werkzeug==2.2.3
gunicorn==20.1.0
requests==2.31.0
pytest==7.4.0



opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-azure-monitor
